+++ webex_assistant_skills_sdk/api/shared/models/__init__.py
@@ -1,0 +1,6 @@
+from webex_assistant_skills_sdk.api.shared.models.rule import DialogueRule
+
+
+__all__ = [
+    DialogueRule,
+]
